♻️ Refactor: 설문 생성 및 결제 기능 로딩 페이지에서 분리
refactor: 설문 생성 및 결제 기능 로딩 페이지에서 분리
